Move old endpoint tests to same file
