+++ api/tests/api_gateway/test_api.py
@@ -789,122 +789,6 @@ def test_remove_all_subscribers_from_list(
     assert data.count() == 1
 
 
-def test_email_list_import(session, list_to_be_updated_fixture, client):
-    data = session.query(Subscription).filter(
-        Subscription.list_id == list_to_be_updated_fixture.id
-    )
-    assert data.count() == 0
-
-    # create email list payload
-    email_list = [f"email{str(x)}@example.com" for x in range(10)]
-    response = client.post(
-        "/listimport",
-        headers={"Authorization": os.environ["API_AUTH_TOKEN"]},
-        json={"list_id": str(list_to_be_updated_fixture.id), "emails": email_list},
-    )
-    assert response.status_code == 200
-    assert response.json() == {"status": "OK"}
-
-    data = session.query(Subscription).filter(
-        Subscription.list_id == list_to_be_updated_fixture.id,
-        Subscription.confirmed.is_(True),
-    )
-    assert data.count() == 10
-
-
-def test_email_list_import_list_id_not_valid(client):
-    # email payload
-    email_list = [f"email{str(x)}@example.com" for x in range(1)]
-    response = client.post(
-        "/listimport",
-        headers={"Authorization": os.environ["API_AUTH_TOKEN"]},
-        json={"list_id": "invalid_list_id", "emails": email_list},
-    )
-    assert response.status_code == 422
-    assert response.json() == {
-        "detail": [
-            {
-                "loc": ["body", "list_id"],
-                "msg": "value is not a valid uuid",
-                "type": "type_error.uuid",
-            }
-        ]
-    }
-
-
-def test_email_list_import_list_not_found(client):
-    # email payload
-    email_list = [f"email{str(x)}@example.com" for x in range(2)]
-    response = client.post(
-        "/listimport",
-        headers={"Authorization": os.environ["API_AUTH_TOKEN"]},
-        json={"list_id": str(uuid.uuid4()), "emails": email_list},
-    )
-    assert response.status_code == 404
-    assert response.json() == {"error": "list not found"}
-
-
-def test_email_list_import_empty_list_of_emails(list_to_be_updated_fixture, client):
-    response = client.post(
-        "/listimport",
-        json={"list_id": str(list_to_be_updated_fixture.id), "emails": list()},
-        headers={"Authorization": os.environ["API_AUTH_TOKEN"]},
-    )
-    assert response.status_code == 422
-    assert response.json() == {
-        "detail": [
-            {
-                "loc": ["body", "emails"],
-                "msg": "ensure this value has at least 1 items",
-                "type": "value_error.list.min_items",
-                "ctx": {"limit_value": 1},
-            }
-        ]
-    }
-
-
-def test_email_list_import_max_emails_allow_is_10000(
-    list_to_be_updated_fixture, client
-):
-    # emails payload
-    email_list = [f"email{str(x)}@example.com" for x in range(10001)]
-    response = client.post(
-        "/listimport",
-        json={"list_id": str(list_to_be_updated_fixture.id), "emails": email_list},
-        headers={"Authorization": os.environ["API_AUTH_TOKEN"]},
-    )
-    assert response.status_code == 422
-    assert response.json() == {
-        "detail": [
-            {
-                "loc": ["body", "emails"],
-                "msg": "ensure this value has at most 10000 items",
-                "type": "value_error.list.max_items",
-                "ctx": {"limit_value": 10000},
-            }
-        ]
-    }
-
-
-@patch("api_gateway.api.db_session")
-def test_email_list_with_unknown_error(
-    mock_db_session, list_to_be_updated_fixture, client
-):
-    mock_session = MagicMock()
-    mock_session.commit.side_effect = SQLAlchemyError()
-    mock_db_session.return_value = mock_session
-    email_list = [f"email{str(x)}@example.com" for x in range(1)]
-    response = client.post(
-        "/listimport",
-        headers={"Authorization": os.environ["API_AUTH_TOKEN"]},
-        json={
-            "list_id": str(list_to_be_updated_fixture.id),
-            "emails": email_list,
-        },
-    )
-    assert response.status_code == 500
-
-
 def test_return_list_subscriber_count(list_with_subscribers, client, session):
     response = client.get("/lists")
     assert response.status_code == 200
